--- src/types/index.ts
@@ -1,7 +1,0 @@
-export interface GlobalSettings {
-    [key: string]: { [key: string]: any };
-}
-
-export interface BlockUtils {
-    isAmpRequest: boolean;
-}
