+++ types/index.ts
@@ -1,0 +1,11 @@
+import { VolComponentButton } from '../src/ButtonComponent/types';
+
+export interface GlobalSettings {
+    globalComponents: {
+        volComponentButton: VolComponentButton;
+    };
+}
+
+export interface BlockUtils {
+    isAmpRequest: boolean;
+}
